+++ Nos/Views/HomeFeedView.swift
@@ -97,8 +97,9 @@ struct HomeFeedView: View {
                                 }
                             }
                             .padding(.horizontal, 15)
+                            .padding(.top, 15)
+                            .padding(.bottom, 0)
                         }
-                        .padding(.top, 15)
                         .readabilityPadding()
                         .id(user.id)
 
@@ -108,7 +109,8 @@ struct HomeFeedView: View {
                                     .padding(.bottom, 15)
                             }
                         }
-                        .padding(.vertical, 15)
+                        .padding(.top, 10)
+                        .padding(.bottom, 15)
                     }
                     .accessibilityIdentifier("home feed")
 
@@ -219,15 +221,17 @@ struct HomeFeedView: View {
 }
 
 fileprivate struct StoryAvatarView: View {
+    var size: CGFloat = 70
+
     var author: Author
     var body: some View {
-        AvatarView(imageUrl: author.profilePhotoURL, size: 54)
-            .padding(.vertical, 10)
-            .background(
+        AvatarView(imageUrl: author.profilePhotoURL, size: size)
+            .padding(1.5)
+            .overlay(alignment: .center) {
                 Circle()
                     .stroke(LinearGradient.diagonalAccent, lineWidth: 3)
-                    .frame(width: 58, height: 58)
-            )
+                    .frame(width: size, height: size)
+            }
     }
 }