+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fixed some cases where a note's contents would never load.
+- Update the color palette.
 
 ## [0.1 (99)] - 2023-12-07Z